Lagt inn støtte for både fullføring av førstegangsbehandling og blankett med vedtak og beregningsdetaljer
